list_view.dart create for seperate list view classes
